TST/REF: collect indexing tests by method
