+++ pandas/tests/indexing/test_loc.py
@@ -1082,6 +1082,182 @@ class TestLocWithMultiIndex:
         tm.assert_series_equal(result, expected)
 
 
+class TestLocSetitemWithExpansion:
+    @pytest.mark.slow
+    def test_loc_setitem_with_expansion_large_dataframe(self):
+        # GH#10692
+        result = DataFrame({"x": range(10 ** 6)}, dtype="int64")
+        result.loc[len(result)] = len(result) + 1
+        expected = DataFrame({"x": range(10 ** 6 + 1)}, dtype="int64")
+        tm.assert_frame_equal(result, expected)
+
+
+class TestLocCallable:
+    def test_frame_loc_getitem_callable(self):
+        # GH#11485
+        df = DataFrame({"A": [1, 2, 3, 4], "B": list("aabb"), "C": [1, 2, 3, 4]})
+        # iloc cannot use boolean Series (see GH3635)
+
+        # return bool indexer
+        res = df.loc[lambda x: x.A > 2]
+        tm.assert_frame_equal(res, df.loc[df.A > 2])
+
+        res = df.loc[lambda x: x.A > 2]
+        tm.assert_frame_equal(res, df.loc[df.A > 2])
+
+        res = df.loc[lambda x: x.A > 2]
+        tm.assert_frame_equal(res, df.loc[df.A > 2])
+
+        res = df.loc[lambda x: x.A > 2]
+        tm.assert_frame_equal(res, df.loc[df.A > 2])
+
+        res = df.loc[lambda x: x.B == "b", :]
+        tm.assert_frame_equal(res, df.loc[df.B == "b", :])
+
+        res = df.loc[lambda x: x.B == "b", :]
+        tm.assert_frame_equal(res, df.loc[df.B == "b", :])
+
+        res = df.loc[lambda x: x.A > 2, lambda x: x.columns == "B"]
+        tm.assert_frame_equal(res, df.loc[df.A > 2, [False, True, False]])
+
+        res = df.loc[lambda x: x.A > 2, lambda x: x.columns == "B"]
+        tm.assert_frame_equal(res, df.loc[df.A > 2, [False, True, False]])
+
+        res = df.loc[lambda x: x.A > 2, lambda x: "B"]
+        tm.assert_series_equal(res, df.loc[df.A > 2, "B"])
+
+        res = df.loc[lambda x: x.A > 2, lambda x: "B"]
+        tm.assert_series_equal(res, df.loc[df.A > 2, "B"])
+
+        res = df.loc[lambda x: x.A > 2, lambda x: ["A", "B"]]
+        tm.assert_frame_equal(res, df.loc[df.A > 2, ["A", "B"]])
+
+        res = df.loc[lambda x: x.A > 2, lambda x: ["A", "B"]]
+        tm.assert_frame_equal(res, df.loc[df.A > 2, ["A", "B"]])
+
+        res = df.loc[lambda x: x.A == 2, lambda x: ["A", "B"]]
+        tm.assert_frame_equal(res, df.loc[df.A == 2, ["A", "B"]])
+
+        res = df.loc[lambda x: x.A == 2, lambda x: ["A", "B"]]
+        tm.assert_frame_equal(res, df.loc[df.A == 2, ["A", "B"]])
+
+        # scalar
+        res = df.loc[lambda x: 1, lambda x: "A"]
+        assert res == df.loc[1, "A"]
+
+        res = df.loc[lambda x: 1, lambda x: "A"]
+        assert res == df.loc[1, "A"]
+
+    def test_frame_loc_getitem_callable_mixture(self):
+        # GH#11485
+        df = DataFrame({"A": [1, 2, 3, 4], "B": list("aabb"), "C": [1, 2, 3, 4]})
+
+        res = df.loc[lambda x: x.A > 2, ["A", "B"]]
+        tm.assert_frame_equal(res, df.loc[df.A > 2, ["A", "B"]])
+
+        res = df.loc[lambda x: x.A > 2, ["A", "B"]]
+        tm.assert_frame_equal(res, df.loc[df.A > 2, ["A", "B"]])
+
+        res = df.loc[[2, 3], lambda x: ["A", "B"]]
+        tm.assert_frame_equal(res, df.loc[[2, 3], ["A", "B"]])
+
+        res = df.loc[[2, 3], lambda x: ["A", "B"]]
+        tm.assert_frame_equal(res, df.loc[[2, 3], ["A", "B"]])
+
+        res = df.loc[3, lambda x: ["A", "B"]]
+        tm.assert_series_equal(res, df.loc[3, ["A", "B"]])
+
+        res = df.loc[3, lambda x: ["A", "B"]]
+        tm.assert_series_equal(res, df.loc[3, ["A", "B"]])
+
+    def test_frame_loc_getitem_callable_labels(self):
+        # GH#11485
+        df = DataFrame({"X": [1, 2, 3, 4], "Y": list("aabb")}, index=list("ABCD"))
+
+        # return label
+        res = df.loc[lambda x: ["A", "C"]]
+        tm.assert_frame_equal(res, df.loc[["A", "C"]])
+
+        res = df.loc[lambda x: ["A", "C"]]
+        tm.assert_frame_equal(res, df.loc[["A", "C"]])
+
+        res = df.loc[lambda x: ["A", "C"], :]
+        tm.assert_frame_equal(res, df.loc[["A", "C"], :])
+
+        res = df.loc[lambda x: ["A", "C"], lambda x: "X"]
+        tm.assert_series_equal(res, df.loc[["A", "C"], "X"])
+
+        res = df.loc[lambda x: ["A", "C"], lambda x: ["X"]]
+        tm.assert_frame_equal(res, df.loc[["A", "C"], ["X"]])
+
+        # mixture
+        res = df.loc[["A", "C"], lambda x: "X"]
+        tm.assert_series_equal(res, df.loc[["A", "C"], "X"])
+
+        res = df.loc[["A", "C"], lambda x: ["X"]]
+        tm.assert_frame_equal(res, df.loc[["A", "C"], ["X"]])
+
+        res = df.loc[lambda x: ["A", "C"], "X"]
+        tm.assert_series_equal(res, df.loc[["A", "C"], "X"])
+
+        res = df.loc[lambda x: ["A", "C"], ["X"]]
+        tm.assert_frame_equal(res, df.loc[["A", "C"], ["X"]])
+
+    def test_frame_loc_setitem_callable(self):
+        # GH#11485
+        df = DataFrame({"X": [1, 2, 3, 4], "Y": list("aabb")}, index=list("ABCD"))
+
+        # return label
+        res = df.copy()
+        res.loc[lambda x: ["A", "C"]] = -20
+        exp = df.copy()
+        exp.loc[["A", "C"]] = -20
+        tm.assert_frame_equal(res, exp)
+
+        res = df.copy()
+        res.loc[lambda x: ["A", "C"], :] = 20
+        exp = df.copy()
+        exp.loc[["A", "C"], :] = 20
+        tm.assert_frame_equal(res, exp)
+
+        res = df.copy()
+        res.loc[lambda x: ["A", "C"], lambda x: "X"] = -1
+        exp = df.copy()
+        exp.loc[["A", "C"], "X"] = -1
+        tm.assert_frame_equal(res, exp)
+
+        res = df.copy()
+        res.loc[lambda x: ["A", "C"], lambda x: ["X"]] = [5, 10]
+        exp = df.copy()
+        exp.loc[["A", "C"], ["X"]] = [5, 10]
+        tm.assert_frame_equal(res, exp)
+
+        # mixture
+        res = df.copy()
+        res.loc[["A", "C"], lambda x: "X"] = np.array([-1, -2])
+        exp = df.copy()
+        exp.loc[["A", "C"], "X"] = np.array([-1, -2])
+        tm.assert_frame_equal(res, exp)
+
+        res = df.copy()
+        res.loc[["A", "C"], lambda x: ["X"]] = 10
+        exp = df.copy()
+        exp.loc[["A", "C"], ["X"]] = 10
+        tm.assert_frame_equal(res, exp)
+
+        res = df.copy()
+        res.loc[lambda x: ["A", "C"], "X"] = -2
+        exp = df.copy()
+        exp.loc[["A", "C"], "X"] = -2
+        tm.assert_frame_equal(res, exp)
+
+        res = df.copy()
+        res.loc[lambda x: ["A", "C"], ["X"]] = -4
+        exp = df.copy()
+        exp.loc[["A", "C"], ["X"]] = -4
+        tm.assert_frame_equal(res, exp)
+
+
 def test_series_loc_getitem_label_list_missing_values():
     # gh-11428
     key = np.array(
